fix : router structure
