+++ package.json
@@ -1,4 +1,7 @@
 {
+  "scripts": {
+    "start": "nodemon --exec ts-node src/server.ts"
+  },
   "devDependencies": {
     "@types/express": "^5.0.3",
     "@types/jsonwebtoken": "^9.0.10",
@@ -6,6 +9,7 @@
     "typescript": "^5.8.3"
   },
   "dependencies": {
+    "cookie-parser": "^1.4.7",
     "express": "^5.1.0",
     "i": "^0.3.7",
     "jsonwebtoken": "^9.0.2"
