+++ src/app.ts
@@ -1,5 +1,4 @@
 import express, { Request, Response } from 'express';
-import { generateToken, authenticateToken } from '././middlewares/auth'; 
 import customHeader from '././middlewares/customHeader';
 import { middleware1, middleware2, middleware3 } from '././middlewares/middlewareChaining';
 import customMiddleware from '././middlewares/custommiddleware';
@@ -26,10 +25,6 @@ app.use(middleware1, middleware2, middleware3);
 
 app.use('/api', routes);
 
-// Root route
-app.get('/', (req: Request, res: Response) => {
-  res.send('Hello World!');
-});
 
 // 404 handler
 app.use((req: Request, res: Response) => {
