+++ src/controllers/other/router.ts
@@ -1,0 +1,10 @@
+import express from 'express'
+import { getHealthCheck,getOtherData,} from './otherControllers';
+
+const router = express.Router();
+
+router.get('/data',getOtherData)
+
+router.get('/heath' , getHealthCheck);
+
+export default router;
